+++ lib/dynatrace_constants.rb
